unify input buffers
